+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Version in progress...
 * "Split annotations by lines" with line thickness 0 px doesn’t work with multiple lines (https://github.com/qupath/qupath/issues/1978)
 * `TransformedServerBuilder.subtractOffset()` can thrown an `IllegalArgumentException` when being serialized (https://github.com/qupath/qupath/issues/1976)
 * Using the lowest resolution image for brightness/contrast settings can be problematic (https://github.com/qupath/qupath/issues/1958)
+* Incorrect error message when merging projects (https://github.com/qupath/qupath/issues/1990)
 
 ### Dependency updates
 * Bio-Formats 8.3.0