+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version in progress...
 * Update to Deep Java Library 0.34.0 (supports PyTorch 2.7.1)
 * *Downgrade* of OpenCV to 4.9.0
   * This is needed for https://github.com/qupath/qupath-extension-align to work
+* The maximum memory preference now uses a percentage, not a number of GB
 
 ### Bug fixes
 * Ensure toolbar button heights are standardized (https://github.com/qupath/qupath/pull/1950)
@@ -20,6 +21,7 @@ Version in progress...
 * `TransformedServerBuilder.subtractOffset()` can thrown an `IllegalArgumentException` when being serialized (https://github.com/qupath/qupath/issues/1976)
 * Using the lowest resolution image for brightness/contrast settings can be problematic (https://github.com/qupath/qupath/issues/1958)
 * Incorrect error message when merging projects (https://github.com/qupath/qupath/issues/1990)
+* Resetting preferences on Windows set max memory to 512 MB (Resetting preferences on Windows set max memory to 512 MB)
 
 ### Dependency updates
 * Bio-Formats 8.3.0